feat: collision(TBD)

+++ package/util/math.go
@@ -1,7 +1,7 @@
 package util
 
 // Clamp keeps a value within a set range
-func Clamp(v, minVal, maxVal int) int {
+func Clamp(v, minVal, maxVal float64) float64 {
 	if v < minVal {
 		return minVal
 	}
